+++ lib/src/material/transparent.dart
@@ -1,0 +1,10 @@
+import 'package:eo_color/palettes.dart';
+
+/// Material Design transparent.
+///
+/// See also
+/// - [transparent](https://api.flutter.dev/flutter/material/Colors/transparent-constant.html)
+class Transparent extends PaletteRGB {
+  /// Fully transparent (invisible).
+  const Transparent() : super(alpha: 0x00, rgb: 0x000000);
+}