+++ CHANGELOG.md
@@ -6,10 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Primaries swatch completion - [#59](https://github.com/dartoos-dev/eo_color/issues/59)
+
 ### Added
 - Red, Pink palettes and swatches - [#57](https://github.com/dartoos-dev/eo_color/issues/57)
   - **Palettes:** Pink, PinkAccent; Red, RedAccent.
   - **Swatches:** Pinks, PinkAccents; Reds, RedAccents.
+- Transparent
 
 ## [0.0.9] - 2021-05-21
 ### Added