+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.10] - 2021-05-22
 ### Changed
 - Primaries swatch completion - [#59](https://github.com/dartoos-dev/eo_color/issues/59)